requirements.txt
psycopg2-binary==2.9.9
pymongo==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
